Fix compiler warnings with clang 4.2 on OS X 10.7.5
On OS X, /usr/include/sys/termios.h typedefs speed_t as unsigned long.
On Linux, it is an unsigned int. This causes printf() format string warnings.
Fix this by down-casting to an unsigned int and using "%u" as format.

Also discard a close() result.

+++ driver_evermore.c
@@ -550,7 +550,7 @@ static bool evermore_speed(struct gps_device_t *session,
 {
     /*@ -type @*/
     gpsd_report(session->context->debug, LOG_PROG,
-		"evermore_speed(%u%c%d)\n", speed, parity,
+		"evermore_speed(%u%c%d)\n", (unsigned int)speed, parity,
 		stopbits);
     /* parity and stopbit switching aren't available on this chip */
     if (parity != session->gpsdata.dev.parity
